+++ findbugs/src/java/edu/umd/cs/findbugs/ba/type/ExceptionSetFactory.java
@@ -45,7 +45,7 @@ public class ExceptionSetFactory implements Serializable {
 	int getIndexOfType(ObjectType type) {
 		Integer index = typeIndexMap.get(type);
 		if (index == null) {
-			index = (Integer)(getNumTypes());
+			index = getNumTypes();
 			typeList.add(type);
 			typeIndexMap.put(type, index);
 		}
